backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
